feat: add same user definition
